feat: convert navbar to AppShell
