--- frontend/src/components/Navbar.module.css
@@ -1,115 +1,0 @@
-.navbarDesktop {
-  height: 100vh;
-  width: calc(300px * var(--mantine-scale));
-  padding: var(--mantine-spacing-md);
-  display: flex;
-  flex-direction: column;
-  border-right: calc(1px * var(--mantine-scale)) solid;
-  border-color: light-dark(
-    var(--mantine-color-gray-3),
-    var(--mantine-color-dark-4)
-  );
-}
-
-.navbarMainDesktop {
-  flex-grow: 1;
-}
-
-.header {
-  padding-bottom: var(--mantine-spacing-md);
-  margin-bottom: calc(var(--mantine-spacing-md) * 1.5);
-  border-bottom: calc(1px * var(--mantine-scale)) solid;
-  border-color: light-dark(
-    var(--mantine-color-gray-3),
-    var(--mantine-color-dark-4)
-  );
-}
-
-.footer {
-  padding-top: var(--mantine-spacing-md);
-  margin-top: var(--mantine-spacing-md);
-  border-top: calc(1px * var(--mantine-scale)) solid;
-  border-color: light-dark(
-    var(--mantine-color-gray-3),
-    var(--mantine-color-dark-4)
-  );
-}
-
-.linkDesktop {
-  display: flex;
-  align-items: center;
-  text-decoration: none;
-  font-size: var(--mantine-font-size-sm);
-  color: light-dark(var(--mantine-color-gray-7), var(--mantine-color-dark-1));
-  padding: var(--mantine-spacing-xs) var(--mantine-spacing-sm);
-  border-radius: var(--mantine-radius-sm);
-  font-weight: 500;
-  width: 100%;
-}
-
-.linkDesktop:hover {
-  background-color: light-dark(
-    var(--mantine-color-gray-0),
-    var(--mantine-color-dark-6)
-  );
-  color: light-dark(var(--mantine-color-black), var(--mantine-color-white));
-}
-
-.linkDesktop[data-active],
-.linkDesktop[data-active]:hover {
-  background-color: var(--mantine-color-indigo-light);
-  color: var(--mantine-color-indigo-light-color);
-}
-
-.linkDesktop[data-active] .linkIcon,
-.linkDesktop[data-active]:hover .linkIcon {
-  color: var(--mantine-color-indigo-light-color);
-}
-
-.linkIcon {
-  color: light-dark(var(--mantine-color-gray-6), var(--mantine-color-dark-2));
-  margin-right: var(--mantine-spacing-sm);
-  width: calc(25px * var(--mantine-scale));
-  height: calc(25px * var(--mantine-scale));
-}
-
-.navbarMobile {
-  width: calc(80px * var(--mantine-scale));
-  height: 100%;
-  padding: var(--mantine-spacing-md);
-  display: flex;
-  flex-direction: column;
-  border-right: calc(1px * var(--mantine-scale)) solid;
-  border-color: light-dark(
-    var(--mantine-color-gray-3),
-    var(--mantine-color-dark-4)
-  );
-}
-
-.navbarMainMobile {
-  flex: 1;
-  margin-top: calc(50px * var(--mantine-scale));
-}
-
-.linkMobile {
-  width: calc(50px * var(--mantine-scale));
-  height: calc(50px * var(--mantine-scale));
-  border-radius: var(--mantine-radius-md);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  color: light-dark(var(--mantine-color-gray-7), var(--mantine-color-dark-0));
-}
-
-.linkMobile:hover {
-  background-color: light-dark(
-    var(--mantine-color-gray-0),
-    var(--mantine-color-dark-5)
-  );
-}
-
-.linkMobile[data-active],
-.linkMobile[data-active]:hover {
-  background-color: var(--mantine-color-indigo-light);
-  color: var(--mantine-color-indigo-light-color);
-}
